requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pillow>=10.3
openpyxl>=3.1
python-multipart>=0.0.9
google-api-python-client>=2.120
google-auth>=2.29
google-auth-oauthlib>=1.2
